decoder_ffmpeg: check earlier that codec/profile is supported by VAAPI
This will allow us to error out earlier (before codec initialization)
if VAAPI does not support the codec/profile. This tries to prevent
the use of the internal FFmpeg software fallback at runtime so we can
use a properly configured software context instead. This allows us to
set the VAAPI avctx thread_count to 1 (which is safer) and rely on the
software avctx fallback which has a thread_count set to 0 (auto). See
068d3bb62e68e461c284693aeb668f99869b9604.